app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.123.9
graphql-core==3.2.7
h11==0.16.0
idna==3.11
lia-web==0.2.3
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
six==1.17.0
starlette==0.50.0
strawberry-graphql==0.287.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
